+++ doc-examples/src/main/java/arez/doc/examples/at_memoize2/NetworkStatus.java
@@ -8,8 +8,8 @@ import arez.annotations.DepType;
 import arez.annotations.Memoize;
 import arez.annotations.OnActivate;
 import arez.annotations.OnDeactivate;
-import elemental2.dom.DomGlobal;
-import elemental2.dom.EventListener;
+import akasha.Global;
+import akasha.EventListener;
 
 @ArezComponent
 public abstract class NetworkStatus
@@ -21,7 +21,7 @@ public abstract class NetworkStatus
   @Memoize( depType = DepType.AREZ_OR_EXTERNAL )
   public boolean isOnLine()
   {
-    return DomGlobal.navigator.onLine;
+    return Global.navigator().onLine();
   }
 
   @ComputableValueRef