+++ doc-examples/src/main/java/arez/doc/examples/lifecycle/BrowserLocation.java
@@ -1,25 +1,25 @@
 package arez.doc.examples.lifecycle;
 
+import akasha.Global;
+import akasha.HashChangeEvent;
+import akasha.HashChangeEventListener;
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
 import arez.annotations.PostConstruct;
 import arez.annotations.PreDispose;
-import elemental2.dom.DomGlobal;
-import elemental2.dom.Event;
-import elemental2.dom.EventListener;
 import javax.annotation.Nonnull;
 
 @ArezComponent
 public abstract class BrowserLocation
 {
-  private final EventListener _listener = this::onHashChangeEvent;
+  private final HashChangeEventListener _listener = this::onHashChangeEvent;
   //DOC ELIDE START
   //DOC ELIDE END
 
   @PostConstruct
   void postConstruct()
   {
-    DomGlobal.window.addEventListener( "hashchange", _listener, false );
+    Global.addHashchangeListener( _listener, false );
     //DOC ELIDE START
     //DOC ELIDE END
   }
@@ -27,12 +27,12 @@ public abstract class BrowserLocation
   @PreDispose
   void preDispose()
   {
-    DomGlobal.window.removeEventListener( "hashchange", _listener, false );
+    Global.removeHashchangeListener( _listener, false );
   }
 
   //DOC ELIDE START
   @Action
-  void onHashChangeEvent( @Nonnull final Event e )
+  void onHashChangeEvent( @Nonnull final HashChangeEvent e )
   {
   }
   //DOC ELIDE END