+++ doc-examples/src/main/java/arez/doc/examples/observer_error/ObserverErrorHandlerExample.java
@@ -1,7 +1,7 @@
 package arez.doc.examples.observer_error;
 
+import akasha.Console;
 import arez.Arez;
-import elemental2.dom.DomGlobal;
 
 @SuppressWarnings( "CodeBlock2Expr" )
 public class ObserverErrorHandlerExample
@@ -9,7 +9,7 @@ public class ObserverErrorHandlerExample
   public static void main( String[] args )
   {
     Arez.context().addObserverErrorHandler( ( ( observer, error, throwable ) -> {
-      DomGlobal.console.error( error + ": Error occurred", throwable );
+      Console.error( error + ": Error occurred", throwable );
     } ) );
   }
 }