+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Upgrade the `org.realityforge.grim` artifacts to version `0.05`.
 * Upgrade the `au.com.stocksoftware.idea.codestyle` artifact to version `1.17`.
+* Migrate from using Elemental2 to Akasha when interacting with the Browser API.
 
 ### [v0.192](https://github.com/arez/arez/tree/v0.192) (2020-12-19) · [Full Changelog](https://github.com/arez/arez/compare/v0.191...v0.192)